Fixed u-tests
