+++ packages/survey-creator-core/tests/property-grid/property-grid.tests.ts
@@ -48,7 +48,6 @@ import { PropertyGridEditorMatrixRateValues } from "../../src/property-grid/matr
 import { editorLocalization } from "../../src/editorLocalization";
 import { SurveyQuestionEditorDefinition } from "../../src/question-editor/definition";
 import { PropertyGridModelTester, findSetupAction } from "./property-grid.base";
-import { CreatorTester } from "../creator-tester";
 
 test("Check property grid survey options", () => {
   const oldValue = Serializer.findProperty(
@@ -533,7 +532,7 @@ test("surveypages property editor and onCollectionItemAllowingCallback", () => {
     item: Base,
     options: ICollectionItemAllowOperations
   ): void => {
-    if(property.name !== "pages") return;
+    if (property.name !== "pages") return;
     const name = (<any>item).name;
     options.allowDelete = name !== "page2";
     options.allowEdit = name !== "page2";
@@ -2017,7 +2016,7 @@ test("Required properties restore on change to empty value", (): any => {
   var question = new QuestionTextModel("q1");
   question.title = "q1Title";
   var options = new EmptySurveyCreatorOptions();
-  options["survey"] = { getAllQuestions: ()=> [question] };
+  options["survey"] = { getAllQuestions: () => [question] };
   var propertyGrid = new PropertyGridModelTester(question, options);
   var titleQuestion = propertyGrid.survey.getQuestionByName("title") as QuestionTextModel;
   titleQuestion.value = "q1t";
@@ -2830,7 +2829,7 @@ test("PropertyEditor for question name", () => {
   const checkedData = ["Row", "panel", "choice", "Item"];
   const errorText = "Do not use reserved words: \"item\", \"choice\", \"panel\", \"row\".";
   let prevName = question.name;
-  for(let i = 0; i < checkedData.length; i ++) {
+  for (let i = 0; i < checkedData.length; i++) {
     const erroredName = checkedData[i];
     const validName = "q" + (i + 1).toString();
     nameQuestion.value = erroredName;
@@ -2846,7 +2845,7 @@ test("PropertyEditor for question name", () => {
   const panel = new PanelModel("p");
   propertyGrid = new PropertyGridModelTester(panel);
   nameQuestion = propertyGrid.survey.getQuestionByName("name");
-  for(let i = 0; i < checkedData.length; i ++) {
+  for (let i = 0; i < checkedData.length; i++) {
     const erroredName = checkedData[i];
     nameQuestion.value = erroredName;
     expect(nameQuestion.value).toBe(erroredName);
@@ -2881,78 +2880,3 @@ test("Allow to enter one space into question title #4416", () => {
   const titleQuestion2 = <QuestionMatrixDynamicModel>(propertyGrid2.survey.getQuestionByName("title"));
   expect(titleQuestion2.value).toBe(" ");
 });
-
-test("Test event prevent default on typing text update mode, fix undo/redo #4889", () => {
-  const testInput = document.createElement("input");
-  const fakeInput = document.createElement("input");
-  document.body.appendChild(testInput);
-  document.body.appendChild(fakeInput);
-
-  const question = new QuestionTextModel("q");
-  const creator = new CreatorTester();
-  creator.registerShortcut("undo_test", {
-    name: "undo",
-    affectedTab: "designer",
-    hotKey: {
-      ctrlKey: true,
-      keyCode: 90,
-    },
-    macOsHotkey: {
-      keyCode: 90,
-    },
-    execute: () => { log += "->undo"; }
-  });
-  creator.registerShortcut("redo_test", {
-    name: "redo",
-    affectedTab: "designer",
-    hotKey: {
-      ctrlKey: true,
-      keyCode: 89,
-    },
-    macOsHotkey: {
-      keyCode: 89,
-    },
-    execute: () => { log += "->redo"; }
-  });
-  let propertyGrid = new PropertyGridModelTester(question, creator);
-  let questionName_RequiredOnBlur = propertyGrid.survey.getQuestionByName("name");
-  let questionTitle_OnTyping = propertyGrid.survey.getQuestionByName("title");
-
-  const eventCtrlZ = {
-    target: testInput,
-    keyCode: 90,
-    ctrlKey: true,
-    preventDefault: () => { log += "->preventDefaultZ"; }
-  };
-  const eventCtrlY = {
-    target: testInput,
-    keyCode: 89,
-    ctrlKey: true,
-    preventDefault: () => { log += "->preventDefaultY"; }
-  };
-
-  let log = "";
-  questionName_RequiredOnBlur.onKeyDown(eventCtrlZ);
-  expect(log).toBe(""); //, "text ctrl+Z for required property - call built-in undo
-
-  questionName_RequiredOnBlur.onKeyDown(eventCtrlY);
-  expect(log).toBe(""); //, "text ctrl+Y" for required property - call built-in redo
-
-  testInput.value = questionName_RequiredOnBlur.value;
-
-  questionName_RequiredOnBlur.onKeyDown(eventCtrlZ);
-  expect(log).toBe("->undo"); //, "text ctrl+Z for required property - call creator undo, built-in undo will do nothing for the save values
-
-  questionName_RequiredOnBlur.onKeyDown(eventCtrlY);
-  expect(log).toBe("->undo->redo"); //, "text ctrl+Y" for required property - call creator redo, built-in undo will do nothing for the save values
-
-  questionTitle_OnTyping.onKeyDown(eventCtrlZ);
-  expect(log).toBe("->undo->redo->undo->preventDefaultZ"); //, "comment ctrl+Z" for onTyping property
-
-  questionTitle_OnTyping.onKeyDown(eventCtrlY);
-  expect(log).toBe("->undo->redo->undo->preventDefaultZ->redo->preventDefaultY"); //, "comment ctrl+Y" for onTyping property
-
-  testInput.remove();
-  fakeInput.remove();
-});
-
